parse endpoint should return modified json
