+++ server/handler/export.go
@@ -80,7 +80,7 @@ func Export(db service.SQLDB) http.HandlerFunc {
 						// DatabaseTypeName JSON is used for arrays of uast nodes and
 						// arrays of strings, but we don't know the exact type.
 						// We try with arry of uast nodes first and any JSON later
-						nodes, err := unmarshallUAST(val)
+						nodes, err := service.UnmarshallUAST(val)
 						if err == nil {
 							b, err := json.Marshal(nodes)
 							if err != nil {
